requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic==2.9.1
python-dotenv==1.0.1
requests==2.32.3
openai==1.40.0
httpx==0.27.2
PyPDF2>=3.0.1
python-multipart>=0.0.5
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4